+++ endpoints-examples/src/main/java/ai/skymind/skil/examples/Main.java
@@ -1,17 +1,21 @@
 package ai.skymind.skil.examples;
 
-import ai.skymind.skil.examples.endpoints.auth.Login;
-import ai.skymind.skil.examples.endpoints.deployments.Deployment;
+import ai.skymind.skil.examples.endpoints.Authorization;
+import ai.skymind.skil.examples.endpoints.Deployment;
 import org.json.JSONArray;
 
 import java.text.MessageFormat;
 
 public class Main {
     public static void main(String[] args) {
-        String authToken = new Login().getAuthToken("admin", "admin");
+        Authorization authorization = new Authorization();
+        String authToken = authorization.getAuthToken("admin", "admin");
         System.out.println(MessageFormat.format("Auth Token: {0}", authToken));
 
-        JSONArray deployments = new Deployment().getAllDeployments();
+        Deployment deployment = new Deployment();
+        JSONArray deployments = deployment.getAllDeployments();
         System.out.println(deployments);
+
+
     }
 }
